+++ shared_web/flask_app.py
@@ -1,49 +1,25 @@
-import logging
 import os
 import subprocess
 import urllib
 from typing import Any, Dict, Optional, Tuple, Union
 
-import sentry_sdk
 from flask import (Blueprint, Flask, Request, Response, redirect, request, send_from_directory,
                    session, url_for)
 from flask.helpers import make_response
 from flask_babel import Babel
 from flask_restx import Api
 from github.GithubException import GithubException
-from sentry_sdk.integrations.flask import FlaskIntegration
 from werkzeug import exceptions, wrappers
 
-from shared import configuration, logger, repo
+from shared import configuration, logger, repo, sentry
 from shared.pd_exception import DoesNotExistException
 
 from . import api, localization, oauth
 from .api import generate_error, return_json
 from .views import InternalServerError, NotFound, Unauthorized
 
+sentry.init()
 
-def sentry_filter(event, hint):  # type: ignore
-    if 'exc_info' in hint:
-        exc_type, exc_value, tb = hint['exc_info']
-        if isinstance(exc_value, OSError):
-            return None
-    return event
-
-
-sentry_token = configuration.get_optional_str('sentry_token')
-if sentry_token:
-    try:
-        sentry_sdk.init(
-            dsn=sentry_token,
-            integrations=[FlaskIntegration()],
-            traces_sample_rate=0.001,
-            before_send=sentry_filter,
-        )
-    except Exception as c:  # pylint: disable=broad-except
-        logging.error(c)
-
-
-# pylint: disable=no-self-use, too-many-public-methods
 class PDFlask(Flask):
     def __init__(self, import_name: str) -> None:
         shared_web_path = os.path.abspath(os.path.dirname(__file__))
